Update and rename script.js to index.js
figured out api's to use. Building foundation of API calls and creating object to hold token data. Also changing file name
